Switch to letting rsync filter our overlaps in docs combined site.
Right now we are doing 3 rsyncs one at a time, but this prevents us
from deleting content that has disappeared from one of the first ones.
So, switch to doing one rsync with multiple sources and let it sort out
files that are now gone to be deleted in the target.
Fixes ticket 7334.