greenwave config: add openQA policies
This is needed for Bodhi to display openQA update test results,
ever since Bodhi was switched from querying resultsdb directly
for relevant results to getting them via greenwave. If we don't
have a policy at all, greenwave doesn't include the results in
the data it passes to Bodhi.

As with the other policies (for Taskotron and the CI pipeline),
this doesn't specify any actual rules for production, so no
gating will be done. I also didn't specify any rules even for
staging, as openQA tests do not run on *all* updates, and there
isn't a greenwave rule for "test muss pass or not be run" (yet).

Signed-off-by: Adam Williamson <awilliam@redhat.com>