mediawiki fedmsg-emit: fix error on creating page (#7415)
This function is set up as a mediawiki hook. It expects to be
passed 12 args. It turns out that mediawiki 1.29.3 - the version
we're running - has two codepaths that call the hook: one passes
it all 12 args, one only passes it 11, missing out the final arg
(undidRevId). @puiterwijk tried to fix this in d85f8c2, but it
seems he misread and set a default value for the 11th arg rather
than the 12th, so it didn't actually fix the problem. Moving the
default value to the 12th arg should fix the problem for good.

Signed-off-by: Adam Williamson <awilliam@redhat.com>