copr: drop 'stage' variable
Since the 'env' by design (see comments in inventory/group_vars/all) can
only contain 'production' or 'staging', we still need to have a way to
to detect 'staging' vs 'development'.

There already was 'devel' variable in copr, but it was set to 'true'
also on staging.  So let's change this.