koji: Fix sending the rpm.sign messages
fedmsg used to automatically serialize datetime object while
fedora-messaging doesn't.
This is fine for code-base one knows but for us on this code base
it has been a bit of a wake-a-mole approach and we were missing
one field (which this commit fixes).
In addition, we're logging the error to the console using python's
logging module, except that it looks like koji isn't quite setting
up handlers for the logs, so they were not showing anywhere :(
(We may need to add print() in the exception handling part in the
future if we run into more problems with this).

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>