postgresql_server / db-koji02: set max_standby_streaming_delay to -1 ( ∞ )
Turns out I had set this on the master (db-koji01) which is ignored.
We need to set it on the standby. With a value of -1, the standby will wait
for conflicting transactions/locks to complete however long it takes.
If this doesn't work to get us a good backup on db-koji02, no harm and we
can try something else.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>