bodhi-backend: re-add fedmsg/base for now
The sync script calls python2 fedmsg stuff. We want to replace it with fedora-messaging,
but we want to do that carefully since mirrormanager depends on messages to decide when
to update directory lists. So, do this for now, and revert it when we have time to manage
the full transition.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>