mirrorlist: do not use containers for mirrorlist-server
The new mirrorlist-server is a single binary and can easily be run as a
service without a container. There have been problems running the
mirrorlist container rootless, especially the combination of running it
as a systemd server and rootless often failed during container
stop/removal.

With this change, starting with Fedora 31, the mirrorlist-server package
is installed and started via systemd to answer haproxy requests. Without
any container.

Signed-off-by: Adrian Reber <adrian@lisas.de>