openqa/dispatcher: handle update test arch config, ppc64le on stg
I enhanced the fedora_openqa job scheduling consumer to have a
setting for which arches update tests should be run on. This
change adds that new setting to the config we write, and enables
ppc64le update tests in staging. If this goes well we will also
enable them in production, and later enable aarch64 as well when
we have more workers.

Signed-off-by: Adam Williamson <awilliam@redhat.com>