Set ansible_ifcfg_whitelist on aarch64-02 and aarch64-03 for now
There is a connected eth2 on each of these machines but no
eth2_ip or eth2_nm in these files, which breaks the 'copy ifcfg
files - non virthost' play. For now I'm just going to set this
whitelist var which should tell ansible to ignore eth2, we can
change the approach later if we want to.

Signed-off-by: Adam Williamson <awilliam@redhat.com>