badges: Add mapping for neuro-sig FAS membership badge
This commit adds a mapping for the `neuro-sig` FAS group to the
`neurofedora-sig-member` badge. This is used in an existing cron job
that automatically awards the badge to members who are already sponsored
into the FAS group, since the automated rule only awards it to people
who are added to the group after the badge is pushed.

Signed-off-by: Justin W. Flory <git@jwf.io>