nagios: change the string checked on the status page
Nagios warns us if status.fp.o isn't running, that's the goal.
But nagios was checking for the presence of the string:
"All systems go".
This is fine, until one system goes down. Nagios tells us about
this system, we go look at it, we (manually) update status.fp.o
so our users know that we know about the outage.
Then nagios tells us that status.fp.o isn't how it should be and
we need to go tell nagios that we know status isn't how it should
be since we updated it ourselves.

So instead of checking for "All systems go" we'll now check for
"Fedora Infrastructure Status" which is at the top of the page
and will remain there as long as status.fp.o is up and regardless
of the state of the rest of the infra.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>