Disable hostkey verification
Since that's run by cron, we have to accept the key.
Given that a MITM of pagure would just enable to intercept
nothing of value (we are using public keys, the data to be commited
can be regenerated trivially and would just concern translation),
disabling StrictHostKeyChecking for that user is IMHO not
a security risk.