fas_client: fix template to correctly apply on pkgs02 and add people02
The ansible_hostname variable is actually the short name of the host,
not the fqdn, so this conditional didn't match before. Switch it to use
startswith and also add people02 as thats the other host people try and
login to often after changing ssh keys.

With this, pkgs02 and people02 should hopefully update ssh keys from fas
every 15min and avoid manual sync requests to the team.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>