Don't do 'nmcli con up' on disabled interfaces
This should avoid us turning on interfaces the config is telling
us shouldn't be turned on. Should only affect openQA-related
boxes, as those are the only ones that use ansible_ifcfg_disabled
at present. It is defined as an empty list in group_vars/all, so
we shouldn't need to worry about it not being defined. Also
converts from the deprecated with_items syntax to loop syntax.

Signed-off-by: Adam Williamson <awilliam@redhat.com>