packager_alias: Another fix for when somehow pagure doesn't return JSON data
From the logs of this cron job running, it appears that every now and then
a request to pagure isn't returning JSON data. Could be because the server
got restarted in the middle or something else.
The result is however the same, the request's data can't be converted to
JSON.

This commit is an addition to 0db60354547c154dcec874470eb5e4792c70e63d which
was already taking care of a similar issue but when retrieving project
information, this one is for retrieving groups information.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>