nagios: the nagios.cfg service isn't phx2_internal
The file for nagios.cfg as a service in nagios isn't in git as
phx2_internal and the nagios playbook is failing because the path for
that service file isn't correct.

It's being copied with the other non-phx2-specific service files in the
following play so this change won't skip the file.