copr-be: automatically cleanup VMs (not)started by resalloc
For some reason, starting VMs sometimes fails and VM is kept idling on
in our VMs list, and we are accounted for it.  This script periodically
checks the list of started VMs and terminates those which we are not
interested in.

The script requires, except for python3, also python3-dateutil.