pagure and distgit/pagure: Drop the db-pagure hostname and populate the db used in the inventor host file instead
Basically, instead of relying on an obscure db-pagure variable that
then needs to be specified in the /etc/hosts file.
Just define the pagure_db_host variable in the host's inventory
file so it exists as a variable available in the playbook/role.
This makes things more explicit and easier to debug/tweak as needed.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>