openqa: conditionally skip relvalconsumer and check-compose
We don't really want/need these roles in IAD yet at all, but the
way the plays are right now we're sort of required to do them.
This skips the queues and roles if a key variable isn't defined,
which should let us skip these cleanly for the IAD deployment.

Signed-off-by: Adam Williamson <awilliam@redhat.com>