iad2: add a hosts file for now on wiki01.iad2 to get auth working
The problem is that id.fedoraproject.org resolves to the iad2 versions
in iad2. This is fine, but break oidc which has a talk between the
provider and the requestor, so if you use the phx2 ipsilon, you need
to use it for the entire thing. Will fix this better in dns soon.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>