base / iptables: add iad2 ips to kojibuilder (phx2) section
We need to add this for s390x machines so they can talk to and be
managed by iad2 stuff. phx2 builders should not be affected, and
s390 builders only get the new rules added, so they should keep working
with phx2. We will need to clean this up after the move and remove all
the phx2 stuff.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>