iad2 / proxy101: candidate registry
Right now, proxy101 is what is resolved to internally in iad2 for
candidate-registry.fedoraproject.org. It has haproxy to reach
oci-candidate-registry01.iad2.fedoraproject.org on port 5000 for this,
but that doesn't work currently due to RHIT firewall.

So, for now we add the vpn endpoint to /etc/hosts there so haproxy works
and internal machines can use the candidate registry.

Once we fix the rhit firewall we should remove this.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>