Drop another blacklist use
This is in something we're never going to deploy again anyway so
it feels kind of absurd, but hey. The variable isn't actually
defined anywhere any more (tflink wiped the vars files that set
it) so this was actually technically broken. This technically
makes it less broken so I am okay with it!

Signed-off-by: Adam Williamson <awilliam@redhat.com>