bodhi/backend: clean and fix the owner-sync-pagure script
Turns out we still carried i586 arch in there while it's long gone
in koji.

The list of extra arches needs to passed as a string, not a list
and a space delimited string, not coma. Otherwise koji crashes on
the other end, but this way it seems to work.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>