drop qa_isolated.
We no longer need this as it was put in place when taskotron was going
to run user provided tests. Since the only left in 'qa' is openqa and it
only tests fedora images/updates and is still also in it's own vlan, we
no longer need to reject things in qa from the rest of their same vlan.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>