openqa: drop openvpn role
In phx2 we used a vpn to openqa01 and all the proxies reached it on
that. In IAD2 that network is more isolated and cannot connect to the
vpn, so we just access it from the IAD2 proxies. If this turns out to be
a problem later we can adjust the RHIT firewall, re-enable vpn and
re-add it to all the proxies.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>