roles/awstats/files/combineHttpLogs.sh: accept DATE_STR
This commit tweaks combineHttpLogs.sh so that you can provide an
optional DATE_STR on the CLI to have it run for a date other than the
default (which is currently 2 days ago, despite the comments about it
being somewhere between 5-3 days..). Links in `latest/` are not updated
when running for user-provided dates.

There's some other refactoring involved but AFAICT this doesn't change
its default behavior at all.

This should let us re-run this script manually to combine logs for the
days that we missed during the colo move etc.