koji_hub: put in hotfixed kojira and config for it
The stock 1.21.1 kojira has a number of problems.
It only does deletes rarely as part of the main thread, so if you are
needing to delete a ton of old repos it will never get to them all.
Also, it checks for it's running newrepos at a place in the loop that
causes it to take a lot longer to notice when it's hit the max for
newrepos.

So, this is the 1.21.1 version, and we will check in the diff/hotfix
which is upstream head version + several patches.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>