koji_hub: readd the heavybuilder channel for chromium
This only works currently for aarch64. we have 2 buildhw's that have a
lot of resources for aarch64, except for disk, so they are great fast
heavy builders. All the other arches machines are pretty much all the
same, so this doesn't help on any of them.

We might someday expand this to more arches and packages when it makes
sense to.

In this case it takes a chromium build that takes ~24 hours
down to taking about an hour.

Note that we are currently keeping these in the default channel too so
they can do normal builds most of the time. That might make chromium
wait a bit, but still will end up being a LOT faster.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>