toddlers: symlink the email_overrides.toml file from distgit_bugzilla_sync to toddlers
We currently need this file in both projects and we don't want to
maintain it in both place.
I first tried to import the file using the structure in the ansible
repo but that doesn't work as it doesn't find the template.
So using a symlink seems like a potentially valid solution.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>