Toddlers: Adjust the configuration file for the new format
Toddlers now comes with a default section which allows having settings
there that are set once and used in multiple toddlers.
For example the dist_git_url is now set in the default section while
the flag_ci_pr and flag_commit_build toddlers are using it.

The runner merges the default configuration section with the toddler
specific section before passing the outcome to the consume method of
the toddler, so we know the toddler will have all the info it needs.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>