fpaste.org: drop for now until ownership or at least resolves to new ips
We don't own fpaste.org, in the past the owners just pointed it to our
proxies and we got a cert for it. However, it's still pointing to the
old phx2 proxies right now. So, drop it until it's fixed to point to the
new ones, or given to us as a domain.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>