ipsilon: set the full path to the theme directory
Currently we let ipsilon's magic load the tempate and asset from
that variable and it looks like it's somehow not doing what we want
with the asset since we can't access the fedora-authn-logo and the
login.css of the Fedora theme.
So let's see if using the full path to the directory helps or not.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>