blockerbugs: set another sebool so openid works
Turns out that the httpd_can_network_connect sebool has to be set in
newer OSs (previous was el7) in order for python_openid to be able to
initiate discovery as part of login.

This is a simple change that sets httpd_can_network_connect to 1 for the
blockerbugs role that should only affect machines running the app.