Fix handling of title-less Planet posts
Some platforms can give us a post without a title. When that happens,
Planet's RSS feed title is "$author" instead of "$author.title", which
causes our split to blow up. Instead, we check to see if there is a ':',
and if so split like normal. If not, explicitly set the title to
"(untitled post)" in order for it to still render.

While I'm at it, update a comment so that it reflects what the code is
actually doing.

Fixes #9251

Signed-off-by: Ben Cotton <bcotton@fedoraproject.org>