git/hooks: ensures the revision sent to fedora-messaging are strings not oid objects
The issue was triggered when multiple commits were pushed at once at
which point the list of revisions contained a list of OID objects
that are not JSON serializable but can be converted to text via
str() or by accessing their .hex attribute.
However, when a single commit is pushed, the list of revisions
contains just the hash of the commit pushed and thus is already
a string. So relying on .hex will not work then.
Using str() becomes the solution that works for both scenarios.

Fixes https://pagure.io/fedora-infrastructure/issue/9247

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>