bodhi: attempt to bump the timeout from 30s to 180s
Both openshift and gunicorn have a timeout. We're ran in the past
in situations where bodhi was a little slow to answer and thus
either gunicorn or openshift were returning a 504 error.
With this commit, we hope to increase both timeout values and thus
give bodhi a little more time to answer.

Hopefully this will help fixing:
https://pagure.io/fedora-infrastructure/issue/9320

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>