Bring back openqa-stg01 and (some)workers as openqa-lab01
We decided that since this wasn't really a staging host, we should stop
calling it 'stg' and putting it in stg group. Since it's used for
testing new openqa and edge features, lab seemed appropriate.

This brings up a openqa-lab01 on qvmhost-x86-01 and a power9 and aarch64
workers. More workers to be added as they are installed. Ther will be at
least one more power9 and 2 more aarch64 boxes.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>