ipsilon: create ipsilon01.stg vm
Right now ipsilon is running in openshift, but sadly in order to work
with ipa instead of fas as it's backend it needs to use sssd and pam
stack. That does not work in a container, so move it back to a vm at
least for now until we sort out what we want to do.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>