inventory / db01: increase memory and cpus and adjust postgresql settings
GIve db01 64gb mem and 24cpus. We have them on that virthost and this
host is one of our busiest.

Adjust the postgresql params based on what I now know from adjusting
db-koji01. This should hopefully improve a number of applications
performance.

Note that we need a restart/reboot to have the cpu changes and postgres
changes take effect. The memory change is already live.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>