proxies: stg.release-monitoring.org make this it's own site
Normally we can just define the site once with it's prod name and use
that in stg with an alias. This works because we have a wildcard ssl
cert for *.fedoraproject.org. So, stg.fedoraproject.org and
fedoraproject.org both work fine. We can't do this for
release-monitoring tho as we use letsencrypt certs and don't have a
wildcard. We could expand out letsencrypt role to get altnames, but just
making these seperate sites should fix it.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>