Add new fedora_nightlies role on openQA prod server
This is to get it off my personal domain (finally). Having it on
the openQA server is still a bit of a hack, but at least it's
under fedoraproject.org now and off my hands. This is a consumer
that runs a static site generator which produces a page full of
links to the most recent nightly compose images, basically.

Signed-off-by: Adam Williamson <awilliam@redhat.com>