Revise Greenwave config again
This revises the Greenwave config again, aiming to be as clear
and efficient as possible. The major practical change here is
we use fedora-* for the product_versions in the 'null' policy
(wildcards like this have been allowed since Greenwave commit
be9a5427 in 2018), and we use lists of decision_contexts instead
of a single decision_context per policy (this has been allowed
since 1.6.0), allowing us to combine two pairs of policies into
single policies. I also revised the comments, policy names and
file organization, taking advantage of these simplifications.

We could make this even more efficient if policies could apply
to multiple subject_types as well, but apparently they can't.
I've filed https://pagure.io/greenwave/issue/609 for that.

Using a wildcard for product_versions in the null policy will
have the practical consequence that whenever a new Fedora release
or dist shows up, if no policy has been prepared for it in advance,
queries for it that match the other elements of the null policy
will succeed rather than failing because no policy is found. i.e.,
practically speaking, Bodhi push queries will succeed rather than
failing. I think this is what we want (since whenever that happens
we just go ahead and add the identifier to the null policy), but
it's worth noting.

Signed-off-by: Adam Williamson <awilliam@redhat.com>