koji_builder: enable bootstrap mode
We need this because f33 rpm switches to sqlite.
So, fedora builds are fine, but epel builds install the buildroot
with f33 rpm/sqlite and then the epel rpm/yum don't know what to do.
This causes mock to make a bootstrap chroot made with the rpm/yum/dnf
of the target and use that to build everyhing after with.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>