Update rsync allows for download servers for tu-chemnitz.de hosts
Currently the rsync control for various the tier0 download servers are
controlled by inventory/group_vars/download. These hosts are allowed
to get downloads days in advance. Control is done by adding them to a
template in the rsync file and controlled by an inventory file for the
download group. [TODO: this is obscure and needs a rethink. It also
uses host names versus ip addresses so we end up with changes like
this one where the reverse DNS name changed.]

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>