greenwave policy: bodhi_update can't use remoterule
So, I noticed today that Bodhi isn't showing openQA results any
more. After poking through greenwave code all morning, this is
my best cut at fixing it, though I still have questions.

The problem seems to be that Greenwave is deciding no policy
'matches' for the subject_type 'bodhi_update'. Because of that
it doesn't attach the test results for that subject type from
ResultsDB to the decision it sends Bodhi (the loop over subjects
is short-circuited before the test results are retrieved if no
policy is considered to 'match').

It decides no policy matches because the only policy we have is
a RemoteRule policy, but Greenwave doesn't think the bodhi_update
subject type "supports" remote rules, and based on various bits
of logic in greenwave, a policy which *only* specifies a
RemoteRule, for a subject type which doesn't "support" them, will
never be considered to "match".

Signed-off-by: Adam Williamson <awilliam@redhat.com>